Extract common symbol prim code
